Move segment deletion cleanup into PinotHelixResourceManager
